+++ misskey-api/src/endpoint/email_address.rs
@@ -1,1 +1,3 @@
+// misskey-dev/misskey#8404
+#[cfg(not(feature = "12-108-0"))]
 pub mod available;