+++ misskey-util/src/client.rs
@@ -392,10 +392,14 @@ pub trait ClientExt: Client + Sync {
     fn mute(&self, user: impl EntityRef<User>) -> BoxFuture<Result<(), Error<Self::Error>>> {
         let user_id = user.entity_ref();
         Box::pin(async move {
-            self.request(endpoint::mute::create::Request { user_id })
-                .await
-                .map_err(Error::Client)?
-                .into_result()?;
+            self.request(endpoint::mute::create::Request {
+                user_id,
+                #[cfg(feature = "12-108-0")]
+                expires_at: None,
+            })
+            .await
+            .map_err(Error::Client)?
+            .into_result()?;
             Ok(())
         })
     }
@@ -3336,12 +3340,16 @@ pub trait ClientExt: Client + Sync {
     }
 
     /// Pins the specified page to the profile.
+    #[cfg(not(feature = "12-108-0"))]
+    #[cfg_attr(docsrs, doc(cfg(not(feature = "12-108-0"))))]
     fn pin_page(&self, page: impl EntityRef<Page>) -> BoxFuture<Result<User, Error<Self::Error>>> {
         let page_id = page.entity_ref();
         Box::pin(async move { self.update_me().set_pinned_page(page_id).update().await })
     }
 
     /// Unpins the page from the profile.
+    #[cfg(not(feature = "12-108-0"))]
+    #[cfg_attr(docsrs, doc(cfg(not(feature = "12-108-0"))))]
     fn unpin_page(&self) -> BoxFuture<Result<User, Error<Self::Error>>> {
         Box::pin(async move { self.update_me().delete_pinned_page().update().await })
     }
@@ -3984,7 +3992,7 @@ pub trait ClientExt: Client + Sync {
     /// client
     ///     .update_meta()
     ///     .set_name("The Instance of Saturn")
-    ///     .max_note_text_length(5000)
+    ///     .local_drive_capacity(5000)
     ///     .update()
     ///     .await?;
     /// # Ok(())